Добавлена возможность добавления, редактирования, удаления тестов через админ-панель, рефакторинг

+++ onboarding-telegram-bot-backend/src/main/java/com/ppteam/onboardingtelegrambot/database/TestQuestionRepository.java
@@ -9,5 +9,5 @@ import java.util.List;
 @Repository
 public interface TestQuestionRepository extends CrudRepository<TestQuestion, Long> {
     @EntityGraph(attributePaths = {"answers"})
-    List<TestQuestion> findByTestId(int testId);
+    List<TestQuestion> findByTestId(long testId);
 }
